# docs/swagger.yaml
basePath: /api
definitions:
...
  project-board-api_internal_dto.AttachmentResponse:
    description: File attachment metadata for boards and projects Contains information
      about uploaded files including S3 URL, size, and content type
    properties:
      contentType:
        example: application/pdf
        type: string
      fileName:
        example: document.pdf
        type: string
      fileSize:
        example: 1024000
        type: integer
      fileUrl:
        example: https://s3.amazonaws.com/bucket/file.pdf
        type: string
      id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
      uploadedAt:
        example: "2024-01-15T10:30:00Z"
        type: string
      uploadedBy:
        example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        type: string
    type: object
  project-board-api_internal_dto.BoardDetailResponse:
    description: 'Detailed board response with value-based customFields, participants,
      and comments customFields contains field type as key and value string as value
      (not UUIDs) Example: {"importance": "high", "role": "developer", "stage": "in_progress"}'
    properties:
      assigneeId:
        example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        type: string
      attachments:
        items:
          $ref: '#/definitions/project-board-api_internal_dto.AttachmentResponse'
        type: array
      authorId:
        example: b2c3d4e5-f6a7-8901-bcde-f12345678901
        type: string
      boardId:
        example: 1275eac5-f0f9-4bee-8235-576a0042f42b
        type: string
      comments:
        items:
          $ref: '#/definitions/project-board-api_internal_dto.CommentResponse'
        type: array
      content:
        example: Add JWT-based authentication to the API
        type: string
      createdAt:
        example: "2024-01-15T10:30:00Z"
        type: string
      customFields:
        additionalProperties:
          type: string
        example:
          importance: high
        type: object
      dueDate:
        example: "2024-12-31T23:59:59Z"
        type: string
      participantIds:
        example:
        - a1b2c3d4-e5f6-7890-abcd-ef1234567890
        - b2c3d4e5-f6a7-8901-bcde-f12345678901
        items:
          type: string
        type: array
      participants:
        items:
          $ref: '#/definitions/project-board-api_internal_dto.ParticipantResponse'
        type: array
      projectId:
        example: 539167fb-b599-41ba-9ead-344a6d0b3a2f
        type: string
      startDate:
        example: "2024-01-01T00:00:00Z"
        type: string
      title:
        example: Implement user authentication
        type: string
      updatedAt:
        example: "2024-01-15T14:20:00Z"
        type: string
    type: object
...
  project-board-api_internal_dto.BoardResponse:
    description: 'Board response with value-based customFields and participant IDs
      customFields contains field type as key and value string as value (not UUIDs)
      Example: {"importance": "high", "role": "developer", "stage": "in_progress"}
      participantIds contains an array of user IDs who are participants of the board'
    properties:
      assigneeId:
        example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        type: string
      attachments:
        items:
          $ref: '#/definitions/project-board-api_internal_dto.AttachmentResponse'
        type: array
      authorId:
        example: b2c3d4e5-f6a7-8901-bcde-f12345678901
        type: string
      boardId:
        example: 1275eac5-f0f9-4bee-8235-576a0042f42b
        type: string
      content:
        example: Add JWT-based authentication to the API
        type: string
      createdAt:
        example: "2024-01-15T10:30:00Z"
        type: string
      customFields:
        additionalProperties:
          type: string
        example:
          importance: high
        type: object
      dueDate:
        example: "2024-12-31T23:59:59Z"
        type: string
      participantIds:
        example:
        - a1b2c3d4-e5f6-7890-abcd-ef1234567890
        - b2c3d4e5-f6a7-8901-bcde-f12345678901
        items:
          type: string
        type: array
      projectId:
        example: 539167fb-b599-41ba-9ead-344a6d0b3a2f
        type: string
      startDate:
        example: "2024-01-01T00:00:00Z"
        type: string
      title:
        example: Implement user authentication
        type: string
      updatedAt:
        example: "2024-01-15T14:20:00Z"
        type: string
    type: object
  project-board-api_internal_dto.CommentResponse:
    properties:
      attachments:
        items:
          $ref: '#/definitions/project-board-api_internal_dto.AttachmentResponse'
        type: array
      boardId:
        type: string
      commentId:
        type: string
      content:
        type: string
      createdAt:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  project-board-api_internal_dto.CreateBoardRequest:
    description: 'Request body for creating a new board with value-based customFields
      customFields should contain field type as key and value string as value Valid
      field types: stage, role, importance Example values: stage="in_progress", role="developer",
      importance="high" participants is an optional array of user IDs to add as board
      participants (max 50) attachmentIds is an optional array of attachment IDs to
      link to the board'
    properties:
      assigneeId:
        example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        type: string
      attachmentIds:
        example:
        - f47ac10b-58cc-4372-a567-0e02b2c3d479
        items:
          type: string
        type: array
      content:
        example: Add JWT-based authentication to the API
        maxLength: 5000
        type: string
      customFields:
        additionalProperties:
          type: string
        example:
          importance: high
        type: object
      dueDate:
        example: "2024-12-31T23:59:59Z"
        type: string
      participants:
        example:
        - a1b2c3d4-e5f6-7890-abcd-ef1234567890
        - b2c3d4e5-f6a7-8901-bcde-f12345678901
        items:
          type: string
        maxItems: 50
        type: array
      projectId:
        example: 539167fb-b599-41ba-9ead-344a6d0b3a2f
        type: string
      startDate:
        example: "2024-01-01T00:00:00Z"
        type: string
      title:
        example: Implement user authentication
        maxLength: 200
        minLength: 1
        type: string
    required:
    - projectId
    - title
    type: object
...
  project-board-api_internal_dto.MoveBoardRequest:
    properties:
      groupByFieldName:
        example: stage
        type: string
      newFieldValue:
        example: in_progress
        type: string
      projectId:
        example: 539167fb-b599-41ba-9ead-344a6d0b3a2f
        type: string
    required:
    - groupByFieldName
    - projectId
    type: object
  project-board-api_internal_dto.MoveBoardResponse:
    properties:
      boardId:
        type: string
      message:
        type: string
      newFieldValue:
        type: string
    type: object
...
  project-board-api_internal_dto.ParticipantResponse:
    description: Participant information for a board
    properties:
      boardId:
        example: 1275eac5-f0f9-4bee-8235-576a0042f42b
        type: string
      createdAt:
        example: "2024-01-15T10:30:00Z"
        type: string
      id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
      userId:
        example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        type: string
    type: object
...
  project-board-api_internal_dto.UpdateBoardRequest:
    description: 'Request body for updating a board with value-based customFields
      customFields should contain field type as key and value string as value Valid
      field types: stage, role, importance Example values: stage="completed", role="designer",
      importance="medium" attachmentIds is an optional array of attachment IDs to
      add to the board'
    properties:
      assigneeId:
        example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        type: string
      attachmentIds:
        example:
        - f47ac10b-58cc-4372-a567-0e02b2c3d479
        items:
          type: string
        type: array
      content:
        example: Refactor JWT implementation
        maxLength: 5000
        type: string
      customFields:
        additionalProperties:
          type: string
        example:
          importance: medium
        type: object
      dueDate:
        example: "2024-12-31T23:59:59Z"
        type: string
      startDate:
        example: "2024-01-01T00:00:00Z"
        type: string
      title:
        example: Update user authentication
        maxLength: 200
        minLength: 1
        type: string
    type: object
...
  project-board-api_internal_response.ErrorResponse:
    properties:
      error: {}
      requestId:
        type: string
    type: object
  project-board-api_internal_response.SuccessResponse:
    properties:
      data: {}
      requestId:
        type: string
    type: object
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    프로젝트 보드 관리 시스템 API 서버입니다.
    Board, Project, Comment, Participant 관리 기능을 제공합니다.
    ALB path-based routing with /api/boards prefix
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Project Board Management API
  version: "1.0"
paths:
...
  /boards:
    get:
      description: |-
        특정 Project에 속한 모든 Board를 조회합니다. 프론트엔드 호환용 엔드포인트
        응답의 customFields는 value 기반 (UUID가 아닌 문자열 값)
        예시: {"importance": "high", "role": "developer", "stage": "in_progress"}
        각 보드는 participantIds (참여자 ID 배열)와 attachments (첨부파일 메타데이터 배열)를 포함합니다
        startDate와 dueDate는 설정된 경우에만 포함됩니다
      parameters:
      - description: Project ID (UUID)
        in: query
        name: projectId
        required: true
        type: string
      - description: 'Custom Fields 필터 JSON 객체. 예시: {\'
        in: query
        name: customFields
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Board 목록 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/project-board-api_internal_dto.BoardResponse'
                  type: array
              type: object
        "400":
          description: 잘못된 Project ID 또는 필터 파라미터
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Project를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Project의 Board 목록 조회 (쿼리 파라미터 방식)
      tags:
      - boards
    post:
      consumes:
      - application/json
      description: |-
        새로운 Board를 생성합니다
        customFields는 value 기반 인터페이스를 사용합니다 (UUID 아님)
        유효한 필드 타입: stage, role, importance
        예시 값: stage="in_progress", role="developer", importance="high"
        잘못된 field value 제공 시 400 에러 반환
        assigneeId가 제공되지 않으면 자동으로 authorId로 설정됩니다
        startDate와 dueDate는 선택 사항이며, startDate는 dueDate보다 이전이어야 합니다
        participants는 선택 사항이며, Board 생성 시 참여자를 함께 추가할 수 있습니다
        participants는 최대 50개의 UUID 배열이며, 중복된 ID는 자동으로 제거됩니다
        participants 추가가 실패해도 Board 생성은 성공하며, 성공한 참여자만 응답에 포함됩니다
        응답의 participantIds 필드에 생성된 참여자 ID 목록이 포함됩니다
        예시: {"participants": ["550e8400-e29b-41d4-a716-446655440001", "550e8400-e29b-41d4-a716-446655440002"]}
      parameters:
      - description: Board 생성 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project-board-api_internal_dto.CreateBoardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Board 생성 성공 (participantIds 포함)
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.BoardResponse'
              type: object
        "400":
          description: 잘못된 요청 또는 유효하지 않은 field value
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Project를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Board 생성
      tags:
      - boards
  /boards/{boardId}:
    delete:
      description: Board를 소프트 삭제합니다
      parameters:
      - description: Board ID (UUID)
        in: path
        name: boardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Board 삭제 성공
          schema:
            $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
        "400":
          description: 잘못된 Board ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Board를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Board 삭제
      tags:
      - boards
    get:
      description: |-
        Board ID로 상세 정보를 조회합니다 (참여자, 댓글, 첨부파일 포함)
        응답의 customFields는 value 기반 (UUID가 아닌 문자열 값)
        예시: {"importance": "high", "role": "developer", "stage": "in_progress"}
        participantIds는 보드에 참여하는 사용자 ID 배열입니다
        attachments는 보드에 첨부된 파일 메타데이터 배열입니다
        startDate와 dueDate는 설정된 경우에만 포함됩니다
      parameters:
      - description: Board ID (UUID)
        in: path
        name: boardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Board 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.BoardDetailResponse'
              type: object
        "400":
          description: 잘못된 Board ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Board를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Board 상세 조회
      tags:
      - boards
    put:
      consumes:
      - application/json
      description: |-
        Board 정보를 수정합니다 (제목, 내용, 단계, 중요도, 역할, 담당자, 날짜)
        customFields는 value 기반 인터페이스를 사용합니다 (UUID 아님)
        유효한 필드 타입: stage, role, importance
        예시 값: stage="completed", role="designer", importance="medium"
        잘못된 field value 제공 시 400 에러 반환
        startDate와 dueDate를 수정할 수 있으며, startDate는 dueDate보다 이전이어야 합니다
      parameters:
      - description: Board ID (UUID)
        in: path
        name: boardId
        required: true
        type: string
      - description: Board 수정 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project-board-api_internal_dto.UpdateBoardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Board 수정 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.BoardResponse'
              type: object
        "400":
          description: 잘못된 요청 또는 유효하지 않은 field value
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Board를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Board 수정
      tags:
      - boards
...
  /boards/{boardId}/move:
    put:
      consumes:
      - application/json
      description: |-
        Board를 다른 컬럼으로 이동합니다. WebSocket을 통해 실시간으로 다른 클라이언트에게 전파됩니다
        groupByFieldName에 해당하는 필드의 값을 newFieldValue로 변경합니다
      parameters:
      - description: Board ID (UUID)
        in: path
        name: boardId
        required: true
        type: string
      - description: Board 이동 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project-board-api_internal_dto.MoveBoardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Board 이동 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.MoveBoardResponse'
              type: object
        "400":
          description: 잘못된 요청
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Board를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Board 이동 (실시간 동기화)
      tags:
      - boards
  /boards/project/{projectId}:
    get:
      description: |-
        특정 Project에 속한 모든 Board를 조회합니다. customFields 파라미터로 필터링 가능 (JSON 형식)
        응답의 customFields는 value 기반 (UUID가 아닌 문자열 값)
        예시: {"importance": "high", "role": "developer", "stage": "in_progress"}
        각 보드는 participantIds (참여자 ID 배열)와 attachments (첨부파일 메타데이터 배열)를 포함합니다
        startDate와 dueDate는 설정된 경우에만 포함됩니다
      parameters:
      - description: Project ID (UUID)
        in: path
        name: projectId
        required: true
        type: string
      - description: 'Custom Fields 필터 JSON 객체. 예시: {\'
        in: query
        name: customFields
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Board 목록 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/project-board-api_internal_dto.BoardResponse'
                  type: array
              type: object
        "400":
          description: 잘못된 Project ID 또는 필터 파라미터
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Project를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Project의 Board 목록 조회
      tags:
      - boards
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT 토큰을 입력하세요. 형식: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"